refactor: dropdown 로직 수정
